+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased][]
 
+### Removed
+
+-   Lodash as dependancy
+
 ## [3.0.0][] - 2021-01-14
 
 ### Changed
@@ -14,6 +18,4 @@
 -   **Node 4 support, lowest version is 10**
 -   Precision option
 
-[unreleased]:
-	https://github.com/niksy/node-sass-json-functions/compare/v3.0.0...HEAD
 [3.0.0]: https://github.com/niksy/node-sass-json-functions/tree/v3.0.0